<Table>: improve reactivity; internalize state
