+++ src/components/Table.svelte
@@ -1,15 +1,271 @@
+<script module lang="ts">
+	// Types
+	import type { Snippet } from 'svelte'
+
+	export type Column<
+		_RowValue = any,
+		_CellValue = any,
+		_ColumnId extends string = string,
+	> = {
+		id: _ColumnId
+		name: string
+		value: (row: _RowValue) => _CellValue
+
+		isSticky?: boolean
+
+		sort?: {
+			isDefault?: boolean
+			defaultDirection: SortDirection
+			compare?: (a: _CellValue, b: _CellValue, rowA: _RowValue, rowB: _RowValue) => number
+		}
+
+		HeaderTitle?: Snippet<
+			[
+				{
+					column: Column<_RowValue, _CellValue, _ColumnId>
+				},
+			]
+		>
+
+		Cell?: Snippet<
+			[
+				{
+					row: _RowValue
+					column: Column<_RowValue, _CellValue, _ColumnId>
+					value: _CellValue
+				},
+			]
+		>
+
+		subcolumns?: Column<_RowValue, _CellValue, _ColumnId>[]
+		isDefaultExpanded?: boolean
+	}
+
+	export enum SortDirection {
+		Ascending = 'asc',
+		Descending = 'desc',
+	}
+
+	type SortState<_ColumnId extends string = string> = {
+		columnId: _ColumnId
+		direction: SortDirection
+	}
+
+	import { SvelteMap, SvelteSet } from 'svelte/reactivity'
+
+	// State
+	export class TableState<
+		_RowValue = any,
+		_CellValue = any,
+		_ColumnId extends string = string,
+	> {
+		columns: Column<_RowValue, _CellValue, _ColumnId>[] = $state([])
+
+		#columnsById = $derived(
+			new SvelteMap(
+				this.columns
+					.flatMap(function flatten(column): (typeof column)[] {
+						return [column, ...(column.subcolumns?.flatMap(flatten) ?? [])]
+					})
+					.map(column => [column.id, column]),
+			),
+		)
+
+		#isColumnExpanded = $state(new SvelteSet<_ColumnId>())
+
+		columnsVisible = $derived.by(() => {
+			const getVisibleColumns = (
+				columns: Column<_RowValue, _CellValue, _ColumnId>[],
+			): Column<_RowValue, _CellValue, _ColumnId>[] =>
+				columns.flatMap(column =>
+					column.subcolumns?.length && this.#isColumnExpanded.has(column.id)
+						? getVisibleColumns(column.subcolumns)
+						: [column],
+				)
+
+			return getVisibleColumns(this.columns)
+		})
+
+		#defaultColumnSort?: SortState<_ColumnId>
+
+		sortState?: SortState<_ColumnId> = $state(this.#defaultColumnSort)
+
+		sortedColumn = $derived(
+			this.sortState?.columnId && this.#columnsById.get(this.sortState.columnId),
+		)
+
+		maxHeaderLevel = $derived.by(() => {
+			const getMaxLevel = (columns: Column<_RowValue, _CellValue, _ColumnId>[]): number =>
+				Math.max(
+					1,
+					...columns.map(column =>
+						!column.subcolumns?.length ? 1 : 1 + getMaxLevel(column.subcolumns),
+					),
+				)
+
+			return getMaxLevel(this.columns)
+		})
+
+		rows = $state<_RowValue[]>([])
+
+		rowIsDisabled?: (row: _RowValue, table: TableState<_RowValue, _CellValue, _ColumnId>) => boolean = $state(undefined)
+
+		displaceDisabledRows: boolean = $state(false)
+
+		rowsSorted = $derived.by(() => {
+			if (!this.sortState) {
+				return this.rows
+			}
+
+			const { columnId, direction } = this.sortState
+
+			const column = this.#columnsById.get(columnId)
+
+			return this.rows.toSorted((a, b) => {
+				if (this.displaceDisabledRows && this.rowIsDisabled) {
+					const isRowADisplaced = this.rowIsDisabled(a, this)
+					const isRowBDisplaced = this.rowIsDisabled(b, this)
+
+					if (isRowADisplaced || isRowBDisplaced) {
+						return isRowADisplaced && isRowBDisplaced ? 0 : isRowADisplaced ? 1 : -1
+					}
+				}
+
+				const aVal = column?.value(a)
+				const bVal = column?.value(b)
+
+				return aVal === undefined || aVal === null
+					? direction === SortDirection.Ascending
+						? 1
+						: -1
+					: bVal === undefined || bVal === null
+						? direction === SortDirection.Ascending
+							? -1
+							: 1
+						: column?.sort?.compare
+							? direction === SortDirection.Ascending
+								? column.sort.compare(aVal, bVal, a, b)
+								: column.sort.compare(bVal, aVal, b, a)
+							: typeof aVal === 'string' && typeof bVal === 'string'
+								? direction === SortDirection.Ascending
+									? aVal.localeCompare(bVal)
+									: bVal.localeCompare(aVal)
+								: aVal < bVal
+									? direction === SortDirection.Ascending
+										? -1
+										: 1
+									: aVal > bVal
+										? direction === SortDirection.Ascending
+											? 1
+											: -1
+										: 0
+			})
+		})
+
+		pageSize: number = $state(Infinity)
+
+		currentPage = $state(1)
+
+		rowsVisible = $derived(
+			this.rowsSorted.slice(
+				(this.currentPage - 1) * this.pageSize || 0,
+				((this.currentPage - 1) * this.pageSize || 0) + this.pageSize,
+			),
+		)
+
+		totalPages = $derived(Math.max(1, Math.ceil(this.rows.length / this.pageSize)))
+
+		canGoBack = $derived(this.rows.length > 0 && this.currentPage > 1)
+
+		canGoForward = $derived(this.rows.length > 0 && this.currentPage < this.totalPages)
+
+		constructor({
+			data,
+			columns,
+			pageSize,
+			rowIsDisabled,
+			displaceDisabledRows,
+		}: {
+			data: _RowValue[]
+			columns: Column<_RowValue, _CellValue, _ColumnId>[]
+			pageSize?: number
+			rowIsDisabled?: (row: _RowValue, table: TableState<_RowValue, _CellValue, _ColumnId>) => boolean
+			displaceDisabledRows?: boolean
+		}) {
+			this.rows = [...data]
+			this.columns = columns
+			this.pageSize = pageSize || Infinity
+
+			const defaultSortedColumn = this.columns.find(column => column.sort?.isDefault)
+
+			this.#defaultColumnSort = this.sortState = defaultSortedColumn && {
+				columnId: defaultSortedColumn.id,
+				direction: defaultSortedColumn.sort?.defaultDirection ?? SortDirection.Ascending,
+			}
+
+			this.rowIsDisabled = rowIsDisabled
+			this.displaceDisabledRows = displaceDisabledRows ?? false
+
+			const initializeIsColumnExpanded = (columns: Column<_RowValue, _CellValue, _ColumnId>[]) => {
+				columns.forEach(column => {
+					if (column.isDefaultExpanded)
+						this.#isColumnExpanded.add(column.id)
+
+					if (column.subcolumns?.length)
+						initializeIsColumnExpanded(column.subcolumns)
+				})
+			}
+
+			initializeIsColumnExpanded(columns)
+		}
+
+		toggleColumnSort = (columnId: _ColumnId) => {
+			const column = this.#columnsById.get(columnId)
+
+			if (!column?.sort) {
+				return false
+			}
+
+			this.sortState = (
+				this.sortState?.columnId !== columnId ?
+					{
+						columnId,
+						direction: column.sort.defaultDirection,
+					}
+				: this.sortState?.direction === column.sort.defaultDirection ?
+					{
+						columnId,
+						direction: column.sort.defaultDirection === SortDirection.Ascending ? SortDirection.Descending : SortDirection.Ascending,
+					}
+				:
+					this.#defaultColumnSort
+			)
+
+			return true
+		}
+
+		isColumnExpanded = (columnId: _ColumnId): boolean => this.#isColumnExpanded.has(columnId)
+
+		toggleIsColumnExpanded = (columnId: _ColumnId) => {
+			if (this.#isColumnExpanded.has(columnId))
+				this.#isColumnExpanded.delete(columnId)
+			else
+				this.#isColumnExpanded.add(columnId)
+		}
+	}
+</script>
+
+
 <script lang="ts" generics="
-	_TableState extends TableState,
-	RowId
+	_RowValue,
+	_RowId,
+	_ColumnId extends string,
+	_CellValue
 ">
 	// Types
 	import type { SvelteHTMLElements } from 'svelte/elements'
-	import type { Snippet } from 'svelte'
-	import { type Column,TableState } from '@/components/TableState.svelte'
 
-	type _RowValue = _TableState extends TableState<infer RowValue, any, any> ? RowValue : any
-	type _CellValue = _TableState extends TableState<any, infer CellValue, any> ? CellValue : any
-	type _ColumnId = _TableState extends TableState<any, any, infer ColumnId> ? ColumnId : string
+	type _TableState = TableState<_RowValue, _CellValue, _ColumnId>
 	type _Column = Column<_RowValue, _CellValue, _ColumnId>
 
 
@@ -41,9 +297,9 @@
 
 		rows: _RowValue[]
 
-		rowId?: (row: _RowValue, index: number) => RowId
-		rowIsDisabled?: (row: _RowValue, table: TableState<_RowValue, _CellValue, _ColumnId>) => boolean
-		onRowClick?: (row: _RowValue, rowId?: RowId) => void
+		rowId?: (row: _RowValue, index: number) => _RowId
+		rowIsDisabled?: (row: _RowValue, table: _TableState) => boolean
+		onRowClick?: (row: _RowValue, rowId?: _RowId) => void
 		displaceDisabledRows?: boolean
 
 		columns: _Column[]
@@ -73,6 +329,19 @@
 			displaceDisabledRows,
 		})
 	)
+
+	$effect(() => {
+		table.rows = rows
+	})
+	$effect(() => {
+		table.columns = columns
+	})
+	$effect(() => {
+		table.rowIsDisabled = rowIsDisabled
+	})
+	$effect(() => {
+		table.displaceDisabledRows = displaceDisabledRows
+	})
 
 	$effect(() => {
 		sortedColumn = table.sortedColumn
